backend/requirements.txt
# RL Studio Backend Requirements
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.1
pydantic>=2.10.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pyyaml>=6.0
requests>=2.31.0

# RL Science Libraries
stable-baselines3>=2.2.0
imitation>=0.4.0
gymnasium>=0.29.0
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Code Generation
openai>=1.12.0

# SkyPilot - ML workload orchestration
skypilot[aws,gcp,azure]>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For async HTTP testing

